+++ changelog.md
@@ -1,4 +1,4 @@
-## April 17th 2019 -
+## April 24th 2019 - 0.8.0
 
 - markRead now supports sending a message_id to mark the channel read up to (and including) that specific message
 - added markAllRead client method